refactor: feature based 형태로 폴더 구조 변경
