+++ packages/assignment-4/src/basic/src/features/cart/components/CartItem.js
@@ -1,0 +1,36 @@
+export function createCartHTML() {
+  return `
+    <div class="bg-gray-100 p-8">
+      <div class="max-w-md mx-auto bg-white rounded-xl shadow-md overflow-hidden md:max-w-2xl p-8">
+        <h1 class="text-2xl font-bold mb-4">장바구니</h1>
+        <div id="cart-items"></div>
+        <div id="cart-total" class="text-xl font-bold my-4"></div>
+        <select id="product-select" class="border rounded p-2 mr-2"></select>
+        <button id="add-to-cart" class="bg-blue-500 text-white px-4 py-2 rounded">추가</button>
+      </div>
+    </div>
+  `;
+}
+
+export function createCartItemHTML(product, quantity) {
+  return `
+    <div id="${product.id}" class="flex justify-between items-center mb-2">
+      <span>${product.name} - ${product.price}원 x ${quantity}</span>
+      <div>
+        <button class="quantity-change bg-blue-500 text-white px-2 py-1 rounded mr-1" data-product-id="${product.id}" data-change="-1">-</button>
+        <button class="quantity-change bg-blue-500 text-white px-2 py-1 rounded mr-1" data-product-id="${product.id}" data-change="1">+</button>
+        <button class="remove-item bg-red-500 text-white px-2 py-1 rounded" data-product-id="${product.id}">삭제</button>
+      </div>
+    </div>
+  `;
+}
+
+export function updateCartDisplay(cartTotal, { finalTotal, discountRate }) {
+  cartTotal.textContent = `총액: ${finalTotal}원`;
+  if (discountRate > 0) {
+    const discountSpan = document.createElement('span');
+    discountSpan.className = 'text-green-500 ml-2';
+    discountSpan.textContent = `(${(discountRate * 100).toFixed(1)}% 할인 적용)`;
+    cartTotal.appendChild(discountSpan);
+  }
+}
