+++ packages/assignment-4/src/basic/src/features/cart/components/ProductSelect.js
@@ -1,0 +1,7 @@
+function createProductOption(product) {
+  return `<option value="${product.id}">${product.name} - ${product.price}원</option>`;
+}
+
+export function populateProductSelect(selectElement, products) {
+  selectElement.innerHTML = products.map(createProductOption).join('');
+}
