+++ packages/assignment-4/src/basic/src/shared/constants/product.js
