+++ packages/assignment-4/src/basic/src/shared/utils/element.js
@@ -1,4 +1,4 @@
-export default function getElement(selector) {
+export function getElement(selector) {
   const element = document.querySelector(selector);
   if (!element) {
     throw new Error(`${selector} 요소를 찾을 수 없습니다.`);
